organize directories and imports
